Moved KeyboardEvents to shared
