+++ src/os/shared.rs
@@ -1,0 +1,15 @@
+use ::std::io::stdin;
+use ::termion::event::Event;
+use ::termion::input::TermRead;
+
+pub struct KeyboardEvents;
+
+impl Iterator for KeyboardEvents {
+    type Item = Event;
+    fn next(&mut self) -> Option<Event> {
+        match stdin().events().next() {
+            Some(Ok(ev)) => Some(ev),
+            _ => None,
+        }
+    }
+}
